feat: Add Zettel extension to text editor
Added the Zettel extension to the list of extensions in the UseTextEditor component. This allows users to utilize the Zettel feature in the text editor for better organization and structure.

Removed editorClear function as it was not being utilized in the component. Replaced it with setEditable and setContent functions to adjust editable state and content of the editor.

The Zettel extension enhances the functionality of the text editor and improves the user experience.