feat: Add trashed field to noteTable
Added the `trashed` field as an integer with boolean mode to the `noteTable` in order to keep track of whether a note has been trashed or not. The default value is set to `false` for new notes. This change enhances the functionality of the table by adding a new attribute to handle note trashing.